Refactor main classes + building full jar
