+++ src/main/java/enlarger/ImageType.java
@@ -1,4 +1,4 @@
-package testEclipse;
+package enlarger;
 
 public enum ImageType {
